requirements.txt
tls_client
httpx
tabulate
fake-useragent
PyYAML